restructure: Move Template Renderer to features
